Implement codeGen following the codeGenExpr algo.
A few changes were necessary:
- Introduce a register allocator and use it consistently.
We can NOT use R0 and R1 directly as specified in the spec,
so make sure to not use them.
- Have the codeGenExpr use the stack when needed. This part
might be broken, we didn't test yet,
but it should switch to stack when enough registers are used.
- Add overflow handlers for all the arith operations (ADD, SUB...)
An overflow handler should exist if the type of the operation
is a float.