codegen: refactor AbstractBinaryExpr codegen
This codegen can now output into any arbritrary register. This will be
useful so that a result can be outputted directly into a register,
instead of always outputting it into the max register,
and having to move it later.

This is also needed to make Assign work properly.

For this Decac code:
int a;
a = a + a;

We will be able to go from:
MOV R2, R4
ADD R4, R4
MOV R4, R2

to
ADD R2, R2